% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{coef.occMod}
\alias{coef.occMod}
\title{Extract regression coefficients from occupancy model, i.e., real parameter estimates}
\usage{
\method{coef}{occMod}(object, param = NULL)
}
\arguments{
\item{object}{an object of class \code{occMod} which is the result from fitting a model using the \code{\link{occMod}} function.}

\item{param}{name of model component to extract regression coefficients for}
}
\value{
a data.frame of the estimates and standard errors
}
\description{
Extract regression coefficients from occupancy model, i.e., real parameter estimates
}
\author{
Darryl MacKenzie
}
