% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{createAicTable}
\alias{createAicTable}
\title{Create AIC table}
\usage{
createAicTable(mod.list, use.aicc = FALSE)
}
\arguments{
\item{mod.list}{a list of objects of class \code{occMod}, i.e., the output from
the function \code{\link{occMod}}.}

\item{use.aicc}{logical argument for whether small-sample adjustment should be used
with AIC. Effective sample set to number of units.
... OR... user-defined value for effective sample size.  So, possibilities are:
\itemize{
\item{\code{0 or FALSE}}{: Use AIC}
\item{\code{1 or TRUE}} {: Use number of sites for eff. sample size to calculate AICc}
\item{\code{n}} {: Use \code{n} for eff. sample size to calculate AICc}
}}
}
\value{
Returns a list with the following components:
 \itemize{
  \item{\code{table}}{ - A data frame that includes:
  \itemize{
    \item{the name}
    \item{AIC value}
    \item{relative difference in AIC}
    \item{AIC model weight}
    \item{-2*log-likelihood}
    \item{number of parameters}
    \item{any warnings for each model}
  }
  Table entries are ranked according to AIC.
  }
  \item{\code{models}}{ - List of models supplied in \code{mod.list}, ordered by AIC}
  \item{\code{ess}}{ - Effective sample size for AICc calculation (0=use AIC, >0=eff. samp. size)}
 }
}
\description{
Create an AIC table of results from a list of previously run models
}
\seealso{
\code{\link{occMod}}
}
\author{
Darryl MacKenzie
}
