% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{createPao}
\alias{createPao}
\title{Create a pao data object.}
\usage{
createPao(data, unitcov = NULL, survcov = NULL,
  nsurveyseason = ncol(data), nmethods = 1, frq = rep(1, nrow(data)),
  title = "PRESENCE Analysis", unitnames = paste0("unit", 1:nrow(data)),
  paoname = "pres.pao")
}
\arguments{
\item{data}{data frame containing the detection (=1) and nondetection (=0) data
with observations from a sampling unit along a row, and surveys in the columns.
Missing values are allowed (=\code{NA}).}

\item{unitcov}{data frame for unit-specific covariates (e.g., habitat,
elevation, etc.) with 1 row per unit (so \code{nrow(unitcov) == nrow(data)}).
Each column of the data frame should contain a single covariate, with the
column name used as the covariate name. Covariatecolumn names shoud not contain
spaces, and avoid using all uppercase as  some terms are used by the package
to define common effects used in models, e.g., SURVEY. If there are no
unit-specific covariates then \code{unitcov=NULL}. Note that additional
unit-specific covariates can be supplied when fitting a model with
\code{\link{occMod}}.}

\item{survcov}{data frame for survey-specific covariates (e.g., survey time,
weather conditions, observer, etc.) with 1 row per survey (so
\code{nrow(survcov) == nrow(data)*ncol(data)}). Each column of the data
frame should contain a single covariate, with the column name used as the
covariate name. Covariate/column names shoud not contain spaces, and avoid
using all uppercase as  some terms are used by the package to define common
effects used in models, e.g., SURVEY. If there are no survey-specific
covariates then \code{survcov=NULL}. Note that additional survey-specific
covariates can be supplied when fitting a model with \code{\link{occMod}}.
Note that the format of survey-specific covariates in \code{survcov} is
different from that used in PRESENCE (i.e., long vs flat format respectively.}

\item{nsurveyseason}{vector of the number of surveys per season (e.g., if
three seasons, with 4 surveys per season, \code{nsurveyseason=c(4,4,4)}).
The default assumes that \code{data} is for a single-season or static
occupancy model.}

\item{nmethods}{integer number of methods per survey}

\item{frq}{vector of frequencies of each detection history (optional)}

\item{title}{(optional) a label included in the PRESENCE output files,
of little practical consequence.}

\item{unitnames}{a vector of string containing labels for each sampling
unit. Included in the results as row names for the real parameter estimates
of the \code{\link{occMod}} object.}

\item{paoname}{string containing a name for the temporary pao}
}
\value{
list of class "\code{pao}" containing the data required to fit an
 occupancy model using \code{\link{occMod}}. The list contains the following
 objects:
  \item{nunits}{number of units or sampling units, i.e., \code{nrow(data)}.}
  \item{nsurveys}{total number of surveys, i.e., \code{ncol(data)}.}
  \item{nseasons}{number of seasons of data.}
  \item{det.data}{detection/nondetection data, i.e., \code{data}.}
  \item{unitcov}{see above. If no covariates are supplied a vector of 1's
  named TEMP is added.}
  \item{survcov}{see above. A factor called SURVEY is automatically added to
  allow survey-specific detection probabilities (but equal across units).}
  \item{nsurveyseason}{see above.}
  \item{title}{see above.}
  \item{unitnames}{see above.}
  \item{surveynames}{a vector of strings providing labels for the survey occasions. Of little practical consequence.}
  \item{paoname}{see above. Equals \code{f} for \code{createPao}.}
}
\description{
\code{createPao} Creates a pao data object from R variables.
}
