% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{fitted.occMod}
\alias{fitted.occMod}
\title{Extract fitted values from occupancy model, i.e., real parameter estimates}
\usage{
\method{fitted}{occMod}(object, param = NULL)
}
\arguments{
\item{object}{an object of class \code{occMod} which is the result from fitting a model using the \code{\link{occMod}} function.}

\item{param}{name of real parameter to extract}
}
\value{
a data.frame of the estimates, standard errors and confidence intervals limits
}
\description{
Extract fitted values from occupancy model, i.e., real parameter estimates
}
\author{
Darryl MacKenzie
}
