% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{genpresEV}
\alias{genpresEV}
\title{generate expected value detection histories}
\usage{
genpresEV(N = 100, K = 4, psi = 0.75, p = 0.4, fp = 0, b = 0,
  eps = 0, gam = 0, sps = 4)
}
\arguments{
\item{N}{number of sites}

\item{K}{number of surveys}

\item{psi}{probability of occupancy}

\item{p}{probability of detection (not false positive)}

\item{fp}{probability of false detection}

\item{b}{probability of "sure" detection}

\item{eps}{probability of local extinction between seasons}

\item{gam}{probability of local colonization between seasons}

\item{sps}{number of surveys per season}
}
\value{
list(hst=det. history matrix, frq=vector of frequencies)
}
\description{
\code{genpresEV} generates expected value detection history frequencies for
single or multi-season models (false-positives optional)
}
\note{
This function generates at least 2^K detection-histories, so if K>8, it may take a
      very long time to run.
}
\examples{
   # generate expected value data for standard single-season model (no false positives)
hstfrq=genpresEV(N=1000, K=4, psi=.8, p=.3)  # generate det. hists and freqs
EVpao=createPao(hstfrq$hst,frq=hstfrq$frq)                  #  create pao file

   # generate expected value data for single-season model with false positive detections
hstfrq=genpresEV(N=1000, K=4, psi=.8, p=.3, fp=0.1, b=0.5)  # generate det. hists and freqs
EVpao=createPao(hstfrq$hst,frq=hstfrq$frq)                  #  create pao file

   # generate expected value data for multi-season model (no false positive detections)
hstfrq=genpresEV(N=1000, K=4, psi=.8, p=.3, eps=.2, gam=.1, sps=2)  # generate det. hists and freqs
EVpao=createPao(hstfrq$hst,frq=hstfrq$frq,nsurveyseason=2)   #  create pao file

}
