% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{get_not01_dm}
\alias{get_not01_dm}
\title{get site/survey covarite names which are not all 0 or 1 in the design matrix
 so the 0/1 covariate names (eg., Intercept) can be omitted from the temp.pao file.}
\usage{
get_not01_dm(dm, covlst)
}
\arguments{
\item{dm}{model design matrix.}

\item{covlst}{character vector of site/survey covariate names}
}
\value{
character vector of site/survey covariate names
}
\description{
\code{get_not01_dm} Gets site/survey covariate names which are not all 0 or 1 in design matrix.
}
