% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{importProject}
\alias{importProject}
\title{Import a PRESENCE project into R.}
\usage{
importProject(pa3)
}
\arguments{
\item{pa3}{a string containing the path (if outside of the current working directory)
and file name of the PRESENCE project summary file with the extension .pa3. e.g.,
\code{import.project("bird_project/bird_project.pa3")}}
}
\value{
Returns an object of class \code{aic.table}, the same as the result of
the \code{\link{createAicTable}} function. The only difference being that the
\code{model} component of the stored \code{\link{occMod}} objects = NULL. i.e.,
the model structure is not imported and left undefined.
}
\description{
Imports the results stored in a PRESENCE project folder into R as an \code{aic.table}
object. Note that in this package version the code has only been developed to import
static occupancy/single-season models, and the first and fourth parameterisation of the
dynamic occupancy/multi-season models. Because design matrices in PRESENCE are constructed
manually, and can be done in multiple ways, no attempt has been made to second-guess the
intended model stucture with respect to covariates. Therefore predictions can not be made
from imported models (unless manually reconfigured), although model averaged estimates of
real parameters of the surveyed units can be obtained.
}
\seealso{
\code{\link{createAicTable}},\code{\link{occMod}},
\code{\link{predict.occMod}},\code{\link{modAvg}}
}
\author{
Darryl MacKenzie.
}
\keyword{models}
