% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHtmlFromAicTable.R
\name{makeHtmlFromAicTable}
\alias{makeHtmlFromAicTable}
\title{write html file with links based on AIC table}
\usage{
makeHtmlFromAicTable(table, fprefix = "", htmlfilename = "tmp.html")
}
\arguments{
\item{table}{An AIC table of model results}

\item{fprefix}{optional prefix for model filenames}

\item{htmlfilename}{name of html file to write (default='tmp.html')}
}
\value{
number of lines written to file.
}
\description{
\code{writeHtmlFromAicTable} Writes html file with links based on AIC table
}
\examples{
\dontrun{
i=makeHtmlFromAicTable(aictable$table)
}

}
\seealso{
\code{\link{createAicTable}}
}
\author{
Jim Hines
}
