% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{modAvg}
\alias{modAvg}
\title{Calculate model averaged estimates}
\usage{
modAvg(aic.tab, param = NULL, index = 1:nrow(aic.tab$table), conf = 0.95,
  predict = FALSE, newdata)
}
\arguments{
\item{aic.tab}{an \code{aicTable} object; the result of \code{\link{createAicTable}}.}

\item{param}{a string with the name of the real parameter to model average, e.g.,
\code{param = "psi"}.  Presently, only works for parameters "psi" and "p".}

\item{index}{a vector of integer values indexing which ranked models are to be model
averaged across. Can be used if only model averaging across a subset of models.
Note that AIC weights are recalcuated such that their sum = 1 for the subset.}

\item{conf}{level for confidence interval (may be vector valued).}

\item{predict}{if \code{FALSE} (default) then model averages observed real estimates.
Otherwise model averages predicted values for \code{param} from the supplied \code{newdata}.}

\item{newdata}{a data frame containing the values of the covariates for which
predictions are required. Variable names must exactly match those used in the data.
May contain varaibles that are not used in some models.}
}
\value{
returns a data frame containing the model averaged estimates, standard errors
and limts of a 95\% confidence interval.
}
\description{
Calculates model averaged estimates of real parameters stored in a \code{aicTable} object;
the result of \code{\link{createAicTable}}. Does not perform model averaging of the
regression coefficients, or beta parameters. Can be used to model average predicted
probabilities too.
}
\seealso{
\code{\link{createAicTable}}, \code{\link{predict.occMod}}
}
\author{
Darryl MacKenzie
}
\keyword{htest}
