% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{new_simplify_dm}
\alias{new_simplify_dm}
\title{simplify design matrix by replacing all individual covariates which are all one or zero
 with '1' or '0' in the design matrix.  This makes the design matrix much easier to read
 and faster for the model.
\code{simplify_dm} Replaces ind. covariates which are all 0 or 1, with 0 or 1.}
\usage{
new_simplify_dm(dm, pao)
}
\arguments{
\item{dm}{model design matrix.}

\item{pao}{pao file created by occMod.???}
}
\value{
simplified design matrix
}
\description{
simplify design matrix by replacing all individual covariates which are all one or zero
 with '1' or '0' in the design matrix.  This makes the design matrix much easier to read
 and faster for the model.
\code{simplify_dm} Replaces ind. covariates which are all 0 or 1, with 0 or 1.
}
