% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.mod.R
\name{occMod}
\alias{occMod}
\title{Fit an occupancy model using Program PRESENCE}
\usage{
occMod(model, cov.list = NULL, data, type = NULL, conf = 0.95,
  modname = NULL, paoname = NULL, outfile = NULL, fixed = NULL,
  initvals = NULL, modfitboot = NULL, VCoutopt = "realbetavc",
  noDerived = F, randinit = 0, maxfn = "32000", quiet = 1,
  limit.real = F, threads = 1, chsq = FALSE, ...)
}
\arguments{
\item{model}{list of formula for each paramter (eg., psi, p, gam, eps...)}

\item{cov.list}{list of covariates used in the model}

\item{data}{pao data object}

\item{type}{a string indicating the type of occupancy model to fit. Currently implemented \code{type} options are:
\tabular{llll}{
  \code{type} \tab Descriptionxxx \tab Function (see for details) \tab Reserved covariate names \cr
  \code{"so"} \tab Static occupancy, single season \tab \code{\link{occMod_SO}}\tab \code{SURVEY} \cr
  \code{"so.fp"} \tab Static occupancy, single season, false-positive detections\tab \code{\link{occMod_SO_fp}}\tab \code{SURVEY} \cr
  \code{"so.mm"} \tab Static occupancy, single season, multi method (or multi-scale) \tab \code{\link{occMod_mm}}\tab \code{SURVEY}\cr
  \code{"so.cd"} \tab Static occupancy, single season, correlated detections \tab \code{\link{occMod_SO_cd}}\tab \code{SURVEY}\cr
  \code{"2sp"}\tab Static or Dynamic occupancy, 2 species, "psiBA" or "nu" param \tab \code{\link{occMod_2SP}}\tab \code{SURVEY,SP,INT,INT_o,INT_d,SEASON}\cr
  \code{"so.2sp.fp"}\tab Static occupancy, single season, 2 species, false positive detections \tab \code{\link{occMod_SO_2SP_cd}}\tab \code{SURVEY,SP,INT,INT_o,INT_d}\cr
  \code{"do.1"} \tab Dynamic occupancy, multi season, 1st param \tab \code{\link{occMod_DO}}\tab \code{SURVEY,SEASON}\cr
  \code{"do.fp"} \tab Dynamic occupancy, multi season, with false-positive detections \tab \code{\link{occMod_DO_fp}}\tab \code{SURVEY,SEASON}\cr
  \code{"do.4"} \tab Dynamic occupancy, multi season, 4th param \tab \code{\link{occMod_DO4}}\tab \code{SURVEY,SEASON}\cr
  \code{"do.ms.1"}\tab Dynamic occupancy, multi season, multi state, 1st param\tab \code{\link{occMod_DO_ms1}}\tab \code{SURVEY,SEASON,DYN,PREV_STATE,STATE}\cr
  \code{"do.ms.2"}\tab Dynamic occupancy, multi season, multi state, 2nd param\tab \code{\link{occMod_DO_ms2}}\tab \code{SURVEY,SEASON,DYN,PREV_STATE,STATE}\cr
  \code{"do.2sp.1"}\tab Dynamic occupancy, multi season, 2 species, psiBA param \tab \code{\link{occMod_2SP}}\tab \code{SURVEY,SP,INT,INT_o,INT_d}\cr
}}

\item{conf}{level for confidence interval (may be vector valued).}

\item{modname}{(optional) a string with the user supplied name for a model.
If \code{NULL}, \code{modname} is created from the formulae supplied in \code{model}.}

\item{paoname}{(optional) a string with the user supplied filename for PRESENCE data and
output files. If \code{NULL}, a generated name is used.}

\item{outfile}{name for output file (use outfile='modname') for outfile named via model name}

\item{fixed}{a data.frame with variables \code{param} and \code{value}. \code{param}
should be a vector of string values containing the names of the real parameters (eg., "p(2)") to fix, and
\code{value} be a numeic vector of the values the real parameter should be fixed at.
 Ordering depends on model \code{type}.}

\item{initvals}{a vector providing initial values for regression coefficients for
the optimization procedure used by PRESENCE. Ordering depends on model \code{type}.}

\item{modfitboot}{number of bootstraps for assessing model fit. Only works for \code{type = "so"}.}

\item{VCoutopt}{option for PRESENCE computation/output of beta/real var-cov matrices}

\item{noDerived}{if TRUE, doesn't print derived estimates from model}

\item{randinit}{number of random initial starting value vectors to try (defalut=0)}

\item{maxfn}{max number of function calls before aborting optimization}

\item{quiet}{0=normal output from Presence,  1=no output (for simulations or repeated runs)}

\item{...}{additional arguments to occMod.XX function.}
}
\value{
Returns a list of class "occMod" and class \code{type}.
  The \code{occMod} object has the following objects:
   \item{modname}{the name of the model}
   \item{model}{a named list containing the right-hand formula for each real parameter type.}
   \item{data}{a copy of the temporary \code{pao} object that is created from the \code{model} argument then used to create the temporary PRESENCE data file for fitting the model.}
   \item{outfile}{name of PRESENCE output file.}
   \item{neg2loglike}{-2*log(likelihood) value for the model evaluated at the maximum likelihood estimates.}
   \item{npar}{number of parameters in the model, calculated as the number of regression coefficients.}
   \item{aic}{value of Akaike's Information Criterion.}
   \item{beta}{list of estimated regression coefficents (sometimes called beta parameters), standard errors and variance-covariance matrix for each parameter type. Exact content depends on the \code{type} of model being fit to the data. Values will be on either the logit or log scale depending on whether the parameter type is a probability or not, respectively.}
   \item{real}{list of real parameter estimates on their natural scale (i.e., 0-1 for probabilities). A data frame for each parameter type the contains the estimate, standard error and limits of the requested confidence intervals. Exact content depends on the \code{type} of model being fit to the data.}
   \item{derived}{list of parameter estimates that are derived from real parameters and not directly estimated. A data frame for each parameter type the contains the estimate, standard error and limits of the requested confidence intervals. Exact content depends on the \code{type} of model being fit to the data.}
   \item{modfit}{list of statistics from parametric bootstrap-based model fit.}
   \item{warnings}{a list containing any warnings from the PRESENCE output file; \code{conv}=number of significant digits convergence was achieved to (=\code{NULL} if >7); and \code{VC}=1 if there is a problem with the variance-covariance matrix (likely non-invertiable observed Hessian matrix), =\code{NULL} otherwise. If \code{conv}<3 then may want to attempt alternative initial values or rescaling covariates to see if different results are obtained. If \code{VC}=1 then all standard errors and confidence intervals should be ignored.}
   \item{version}{a list containing the version numbers for PRESENCE and RPresence used when fitting the model}
}
\description{
\code{occMod} returns a list variable containing detection-history and covariate data.
}
\details{
choices for \code{VCoutopt} are:
\itemize{
 \item{"nose"}{only compute likelihood and beta estimates, no beta var-cov matrix or beta SE's (useful for faster model selection of big models or simulations)}
 \item{"betavc"}{compute but don't print beta var-cov matrix so we get beta SE's, don't compute real params}
 \item{"noreal"}{compute and print beta beta var-cov matrix, don't compute real params (I think this is what you want for RPresence)}
 \item{"novcs"}{compute real params and real var-cov matrix, don't print either var-cov matrix.}
 \item{"realbetavc"}{(default) compute real params and real var-cov matrix, print only beta var-cov matrix.}
 \item{"bothvc"}{print both var-cov matrices.}
}
}
\examples{
# load a PRESENCE data file
filename<-system.file("extdata/weta.pao",package="RPresence")
weta.data<-readPao(filename)

## convert indicator variables to categorical covariates
weta.data$unitcov$Habitat[weta.data$unitcov$Browsed==1] <- "browsed"
weta.data$unitcov$Habitat[weta.data$unitcov$Unbrowsed==1] <- "unbrowsed"
weta.data$survcov$Obs<-as.factor(
     1*weta.data$survcov$Obs1+
     2*weta.data$survcov$Obs2+
     3*weta.data$survcov$Obs3)

## fit some models
mod1<-occMod(model=list(psi~Habitat,p~SURVEY+Obs),data=weta.data,type="so")
mod2<-occMod(model=list(psi~Habitat,p~Obs),data=weta.data,type="so")
mod3<-occMod(model=list(psi~Habitat,p~SURVEY),data=weta.data,type="so")
mod4<-occMod(model=list(psi~Habitat,p~1),data=weta.data,type="so")
mod5<-occMod(model=list(psi~1,p~SURVEY+Obs),data=weta.data,type="so")
mod6<-occMod(model=list(psi~1,p~Obs),data=weta.data,type="so")
mod7<-occMod(model=list(psi~1,p~SURVEY),data=weta.data,type="so")
mod8<-occMod(model=list(psi~1,p~1),data=weta.data,type="so")
#
## create AIC table
models<-list(mod1,mod2,mod3,mod4,mod5,mod6,mod7,mod8)
results<-createAicTable(models)
summary(results)

## perform model averageing
psi.ma<-modAvg(aic.tab=results,param="psi")
p.ma<-modAvg(aic.tab=results,param="p")

\dontrun{
m1<-occMod(model=list(psi~habitat,gamma~habitat+SEASON,epsilon~habitat+SEASON,
                       p~habitat+SEASON),data=data,type="do.1")
}
}
\seealso{
\code{\link{occMod_SO}},\code{\link{occMod_mm}},\code{\link{occMod_2SP}},
         \code{\link{occMod_DO}},\code{\link{occMod_DO4}},\code{\link{occMod_DO_ms2}},\code{\link{occMod_2SPfp}}
}
\author{
Darryl MacKenzie
}
