% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.mod.do.R
\name{occMod_DO}
\alias{occMod_DO}
\title{Fit dynamic (multi-season) occupancy model using the first parameterisation in PRESENCE.}
\usage{
occMod_DO(psi = call(), psi.cov = data$unitcov, gamma = call(),
  gamma.cov = data$unitcov, epsilon = call(), epsilon.cov = data$unitcov,
  p = call(), p.cov = cbind(data$unitcov, data$survcov), theta = NULL,
  theta.cov = data$unitcov, th0pi = NULL, th0pi.cov = data$unitcov,
  modname = NULL, paoname = NULL, outfile, model = 200, fixed = NULL,
  initvals = NULL, data = data, conf = 0.95, miscopts = "")
}
\arguments{
\item{psi}{the right-hand side of the formula for the model to fit for occupancy probability in the first season.}

\item{psi.cov}{a data frame containing the unit-specific covariates to use for the occupancy component of the model, with number of rows = \code{data$nunits}.}

\item{gamma}{the right-hand side of the formula for the model to fit for colonization probability in each season.}

\item{gamma.cov}{a data frame containing the unit-specific covariates to use for the colonization component of the model, with number of rows = \code{data$nunits*(data$nseasons-1)}.}

\item{epsilon}{the right-hand side of the formula for the model to fit for extinction probability in each season.}

\item{epsilon.cov}{a data frame containing the unit-specific covariates to use for the extinction component of the model, with number of rows = \code{data$nunits*(data$nseasons-1)}.}

\item{p}{the right-hand side of the formula for the model to fit for detection probability.}

\item{p.cov}{a data frame containing the survey-specific covariates to use for the detection component of the model.}

\item{theta}{(optional) the right-hand side of the formula for the model to fit for local occupancy parameter (th0,th1). Set equal to NULL for standard dynamic model.}

\item{theta.cov}{a data frame containing the survey-specific covariates to use for the local occupancy component of the model.}

\item{th0pi}{the right-hand side of the formula for the model to fit for initial local occupancy probability.}

\item{th0pi.cov}{a data frame containing the site-specific covariates to use for th0pi.}

\item{modname}{(optional) a string containing the model name}

\item{paoname}{(optional) a string containing the filename for the temporary PRESENCE data file.}

\item{outfile}{name for output file (use outfile='modname') for outfile named via model name}

\item{model}{the PRESENCE model code. DO NOT CHANGE.}

\item{fixed}{a single-column matrix containing values for real parameters to be fixed at. \code{rownnames(fixed)} should contain the index of the real parameters to be fixed.}

\item{initvals}{initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.}

\item{data}{the \code{pao} data object containing the detection data and other information.}

\item{conf}{limits for confidence intervals as a proportion (defalut=0.95 for 95\% conf. interval limits)}

\item{miscopts}{(see \code{\link{occMod}})}
}
\value{
returns a list of class \code{"occMod"} and \code{"do1"}.

\code{occMod$beta} contains the objects:
 \item{psi}{estimated logistic regression coefficients and standard errors for probability of occurrence in the first year.}
 \item{psi.VC}{variance-covariance matrix for logistic regression coefficients for probability of occurrence in the first year.}
 \item{gamma}{estimated logistic regression coefficients and standard errors for probability of colonization.}
 \item{gamma.VC}{variance-covariance matrix for logistic regression coefficients for probability of colonization.}
 \item{epsilon}{estimated logistic regression coefficients and standard errors for probability of extinction.}
 \item{epsilon.VC}{variance-covariance matrix for logistic regression coefficients for probability of extinction.}
 \item{p}{estimated logistic regression coefficients and standard errors for probability of detection.}
 \item{p.VC}{variance-covariance matrix for logistic regression coefficients for probability of detection.}
 \item{VC}{the full variance-covariance matrix for all logistic regression coefficients.}

 \code{occMod$real} contains the objects:
 \item{psi}{estimated probabilities of occurrence for each sampling unit, along with standard errors and limits of 95\% confidence interval. Estimates are provided for the first season (calculated directly from the estimated \code{beta} parameters). Estimates for later seasons are provided as derived parameters. The season for which an estimate applies to can be identified from the final number of the row names (\code{rownames(psi)}).}
 \item{gamma}{estimated probabilities of colonization for each sampling unit, along with standard errors and limits of 95\% confidence interval. The beginning season for which an estimate applies to can be identified from the final number of the row names (\code{rownames(gamma)}).}
 \item{epsilon}{estimated probabilities of extinction for each sampling unit, along with standard errors and limits of 95\% confidence interval. The beginning season for which an estimate applies to can be identified from the final number of the row names (\code{rownames(epsilon)}).}
 \item{phi}{estimated probabilities of persistence (i.e., \code{(1-epsilon)}) for each sampling unit, along with standard errors and limits of 95\% confidence interval. The beginning season for which an estimate applies to can be identified from the final number of the row names (\code{rownames(phi)}).}
 \item{p}{estimated probabilities of detection for each survey, along with standard errors and limits of 95\% confidence interval.}

 \code{occMod$derived} contains the objects:
 \item{psi}{estimated probabilities of occurrence for each sampling unit for second season onwards, along with standard errors and limits of 95\% confidence interval. The season for which an estimate applies to can be identified from the final number of the row names (\code{rownames(psi)}).}
}
\description{
This is not intended for direct use, but instead the \code{\link{occMod}}
       function should be used with \code{type="do.1"}.
}
\details{
Pre-defined covariates:
\itemize{
\item{SURVEY} {categorical covariate indicating survey (for detection parameters)}
\item{SEASON} {categorical covariate indicating season (for detection or col/ext parameters)}
}
}
\seealso{
\code{\link{occMod}}
}
\author{
Darryl MacKenzie
}
