% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.mod.do4.R
\name{occMod_DO4}
\alias{occMod_DO4}
\title{Fit dynamic (multi-season) occupancy model using the fourth parameterisation
(random changes in occupancy, epsilon=1-gamma) in PRESENCE.}
\usage{
occMod_DO4(psi = call(), psi.cov = data$unitcov, p = call(),
  p.cov = cbind(data$unitcov, data$survcov), modname = NULL,
  paoname = NULL, outfile, model = 240, fixed = NULL, initvals = NULL,
  data, conf, miscopts = "")
}
\arguments{
\item{psi}{the right-hand side of the formula for the model to fit for occupancy probability
in each season.}

\item{psi.cov}{a data frame containing the unit-specific covariates to use for the occupancy
component of the model, with number of rows = \code{data$nunits} or \code{data$nunits*data$nseasons}.
If the shorter version of the data frame is supplied the rows are recycled to the longer length.}

\item{p}{the right-hand side of the formula for the model to fit for detection probability.}

\item{p.cov}{a data frame containing the survey-specific covariates to use for the detection component
of the model.}

\item{modname}{(optional) a string containing the model name}

\item{paoname}{(optional) a string containing the filename for the temporary PRESENCE data file.}

\item{outfile}{name for output file (use outfile='modname') for outfile named via model name}

\item{model}{The PRESENCE model code. DO NOT CHANGE.}

\item{fixed}{a single-column matrix containing values for real parameters to be fixed at.
\code{rownnames(fixed)} should contain the index of the real parameters to be fixed.}

\item{initvals}{initial values for the beta parameters at which PRESENCE begins the optimisation.
The default values in PRESENCE is 0.}

\item{data}{the \code{pao} data object containing the detection data and other information.}

\item{conf}{limits for confidence intervals as a proportion (defalut=0.95 for 95\% conf. interval limits)}

\item{miscopts}{see \code{\link{occMod}}}
}
\value{
list of class \code{"occMod"} and \code{"do4"}.

\code{occMod$beta} contains the objects:

 \item{psi}{estimated logistic regression coefficients and standard errors for probabilities of occurrence in each season.}
 \item{psi.VC}{variance-covariance matrix for psi.}
 \item{p}{estimated logistic regression coefficients and standard errors for probabilities of detection in each survey.}
 \item{p.VC}{variance-covariance matrix for p.}
 \item{VC}{the full variance-covariance matrix for all logistic regression coefficients.}

\code{occMod$real} contains the objects:
 \item{psi}{estimated probabilities of occurrence for each sampling unit, in each season,
 along with standard errors and limits of 95\% confidence interval. The season for which an
 estimate applies to can be identified from the final number of the row names
 (\code{rownames(psi)}).}
 \item{p}{estimated probabilities of detection for each survey, along with standard errors
 and limits of 95\% confidence intervals.}

\code{occMod$derived} contains the objects:
 \item{gamma}{estimated probabilities of colonization for each sampling unit, along with
 standard errors and limits of 95\% confidence intervals. The beginning season for which an
 estimate applies to can be identified from the final number of the row names
 (\code{rownames(gamma)}).}
 \item{epsilon}{estimated probabilities of extinction for each sampling unit, along with
 standard errors and limits of 95\% confidence intervals. The beginning season for which an
 estimate applies to can be identified from the final number of the row names
 (\code{rownames(epsilon)}). Under this parameterisation, extinction probabilities have the
 same value as 1-colonization probabilties}

Note: For this parameterization, \eqn{\gamma_t = \psi_{t+1}}{gamma(t)=psi(t+1)}, due to the
constraint, \eqn{\epsilon=1-\gamma}{epsilon=1-gamma}
           \preformatted{
           ie., psi(t+1)=psi(t)*(1-epsilon(t))+(1-psi(t))*gamma(t)
                psi(t+1)=psi(t)*gamma(t)+(1-psi(t))*gamma(t)
                psi(t+1)=psi(t)*gamma(t)+gamma(t)-psi(t)*gamma(t)
                psi(t+1)=gamma(t)}
}
\description{
This is not intended for direct use, but instead the \code{\link{occMod}} function should
be used with \code{type="do.4"}.
}
\seealso{
\code{\link{occMod}}
}
\author{
Darryl MacKenzie
}
