% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.mod.do.fp.R
\name{occMod_DO_fp}
\alias{occMod_DO_fp}
\title{Fit static occupancy/single season model}
\usage{
occMod_DO_fp(psi = call(), psi.cov = data$unitcov, gamma = call(),
  gamma.cov = data$unitcov, epsilon = call(), epsilon.cov = data$unitcov,
  p11 = call(), p11.cov = cbind(data$unitcov, data$survcov), p10 = call(),
  p10.cov = cbind(data$unitcov, data$survcov), b = call(),
  b.cov = cbind(data$unitcov, data$survcov), modname = NULL,
  paoname = NULL, outfile, model = 100, fixed = NULL, initvals = NULL,
  data, conf, miscopts = "")
}
\arguments{
\item{psi}{the right-hand side of the formula for the model to fit for occupancy probability.}

\item{psi.cov}{a data frame containing the unit-specific covariates to use for the occupancy
component of the model.}

\item{gamma}{the right-hand side of the formula for the model to fit for colonization probability in each season.}

\item{gamma.cov}{a data frame containing the unit-specific covariates to use for the colonization component of the model, with number of rows = \code{data$nunits*(data$nseasons-1)}.}

\item{epsilon}{the right-hand side of the formula for the model to fit for extinction probability in each season.}

\item{epsilon.cov}{a data frame containing the unit-specific covariates to use for the extinction component of the model, with number of rows = \code{data$nunits*(data$nseasons-1)}.}

\item{p11}{the right-hand side of the formula for the model to fit for "sure" detection probability.}

\item{p11.cov}{a data frame containing the survey-specific covariates to use for the detection
component of the model.}

\item{p10}{the right-hand side of the formula for the model to fit for false-positive detection probability.}

\item{p10.cov}{a data frame containing the survey-specific covariates to use for the detection
component of the model.}

\item{b}{the right-hand side of the formula for the model to fit for Pr(detection is "sure")}

\item{b.cov}{a data frame containing the survey-specific covariates to use for the b}

\item{modname}{(optional) a string containing the model name}

\item{paoname}{(optional) a string containing the filename for the temporary PRESENCE data file.}

\item{outfile}{name for output file (use outfile='modname') for outfile named via model name}

\item{model}{the PRESENCE model code. DO NOT CHANGE.}

\item{fixed}{a single-column matrix containing values for real parameters to be fixed at.
\code{rownnames(fixed)} should contain the index of the real parameters to be fixed.}

\item{initvals}{initial values for the beta parameters at which PRESENCE begins the optimisation.
The default values in PRESENCE is 0.}

\item{data}{the \code{pao} data object containing the detection data and other information.}

\item{conf}{level for confidence interval (may be vector valued).}

\item{miscopts}{see \code{\link{occMod}}}
}
\value{
returns a list of class \code{"occMod"} and \code{"soFp"}.

\code{occMod$beta} contains the objects:
\item{psi}{estimated logistic regression coefficients and standard errors for probability of occurrence.}
\item{psi.VC}{variance-covariance matrix for logistic regression coefficients for probability of occurrence.}
\item{gamma}{estimated logistic regression coefficients and standard errors for probability of colonization.}
\item{gamma.VC}{variance-covariance matrix for logistic regression coefficients for probability of colonization.}
\item{epsilon}{estimated logistic regression coefficients and standard errors for probability of extinction.}
\item{epsilon.VC}{variance-covariance matrix for logistic regression coefficients for probability of extinction.}
\item{p}{estimated logistic regression coefficients and standard errors for probability of detection.}
\item{p.VC}{variance-covariance matrix for logistic regression coefficients for probability of detection.}
\item{VC}{the full variance-covariance matrix for all logistic regression coefficients.}

\code{occMod$real} contains the objects:
\item{psi}{estimated probabilities of occurrence for each sampling unit, along with standard
errors and limits of 95\% confidence interval.}
\item{p}{estimated probabilities of detection for each survey, along with standard errors and
limits of 95\% confidence interval.}

\code{occMod$derived} contains the objects:
\item{psi_c}{estimated probabilities of occurrence given the detection history for each sampling
unit, along with standard errors and limits of 95\% confidence interval. Will be \code{=1} for
any unit where the species was detected at least once.}
}
\description{
This is not intended for direct use, but instead the \code{\link{occMod}} function should be used with \code{type="do.fp"}.
}
\seealso{
\code{\link{occMod}}, \code{pao}
}
\author{
Darryl MacKenzie
}
