% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.mod.do.ms1.R
\name{occMod_DO_ms1}
\alias{occMod_DO_ms1}
\title{Fit multi-state dynamic (multi-season) occupancy model using the 1st parameterisation (\eqn{\psi/\phi}).}
\usage{
occMod_DO_ms1(psi = call(), psi.cov = data$unitcov, phi = call(),
  phi.cov = data$unitcov, p = call(), p.cov = data$survcov,
  modname = NULL, paoname = NULL, outfile, model = 6000, fixed = NULL,
  initvals = NULL, data, conf, miscopts = "")
}
\arguments{
\item{psi}{the right-hand side of the formula for the model to fit for occupancy probability.}

\item{psi.cov}{a data frame containing the unit-specific covariates to use for the occupancy component of the model.}

\item{phi}{the right-hand side of the formula for the model to fit for occupancy transition probabilities,.}

\item{phi.cov}{a data frame containing the unit-specific covariates to use for the occupancy component of the model.}

\item{p}{the right-hand side of the formula for the model to fit for detection probability.}

\item{p.cov}{a data frame containing the survey-specific covariates to use for the detection component of the model.}

\item{modname}{(optional) a string containing the model name}

\item{paoname}{(optional) a string containing the filename for the temporary PRESENCE data file.}

\item{outfile}{name for output file (use outfile='modname') for outfile named via model name}

\item{model}{the PRESENCE model code. DO NOT CHANGE.}

\item{fixed}{a 2-column matrix containing indices and values for real parameters to be fixed at.}

\item{initvals}{initial values for the beta parameters at which PRESENCE begins the optimisation. Default = 0.}

\item{data}{the \code{pao} data object containing the detection data and other information.}

\item{conf}{level for confidence interval (may be vector valued).}

\item{miscopts}{see \code{\link{occMod}}}
}
\value{
returns a list of class \code{"occMod"} and \code{"do.1"}.

\code{occMod$beta} contains the objects:
 \item{psi}{estimated logistic regression coefficients and standard errors for probability of occurrence in the first year.}
 \item{psi.VC}{variance-covariance matrix for logistic regression coefficients for probability of occurrence in the first year.}
 \item{phi}{estimated logistic regression coefficients and standard errors for probability of state transition between years.}
 \item{phi.VC}{variance-covariance matrix for logistic regression coefficients for probability of transition between years.}
 \item{p}{estimated logistic regression coefficients and standard errors for probability of detection.}
 \item{p.VC}{variance-covariance matrix for logistic regression coefficients for probability of detection.}
 \item{VC}{the full variance-covariance matrix for all logistic regression coefficients.}

 \code{occMod$real} contains the objects:
 \item{psi}{estimated probabilities of occurrence for each sampling unit and state, along with standard errors and limits
 of 95\% confidence interval.}
 \item{phi}{estimated probabilities of transition between occupancy states for each sampling unit and state, along with standard errors and limits
 of 95\% confidence interval.}
 \item{p}{estimated probabilities of detection for each survey, along with standard errors and limits of 95\% confidence interval.}
}
\description{
This is not intended for direct use, but instead the \code{\link{occMod}} function should be used with \code{type="do.ms.1"}.
}
\note{
NOTE THAT THIS FUNCTION HAS NOT BEEN EXTENSIVELY TESTED SO EXACT IMPLEMENTATION MAY CHANGE.

psi built-in covariates:
\itemize{
   \item{STATE   - for model where psi (initial occupancy) depends on occupancy state.}
}
phi built-in covariates:
\itemize{
   \item{FROM - for model where psi depends on state species is transitioning from.}
   \item{TO   - for model where psi depends on state species is transitioning to.}
   \item{FROM_TO - for model where psi depends on "from" state and "to" state (equalivent to "FROM*TO").}
   \item{SEASON - for model where psi changes from season to season.}
}
P built-in covariates:
\itemize{
   \item{OSTATE - detection depends on observed state of species.}
   \item{OTSTATE - detection depends on both observed and true state of species.}
   \item{SURVEY  - detection varies by survey.}
   \item{SEASON  - detection varies by season (constant within season).}
}
Note: P21 (prob detect as state=2, when true state=1) must be fixed to 0, or model will not converge
Also: setting outfile="modname" causes RPresence to save the output file using the model name
                        (with ".out" extension). Otherwise, you can name it something else.
                        Or, set outfile=NULL to not save the output file.
}
\seealso{
\code{\link{occMod}}, \code{\link{occMod_DO_ms2}}
}
\author{
Jim Hines
}
