% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.mod.do.ms2.R
\name{occMod_DO_ms2}
\alias{occMod_DO_ms2}
\title{Fit multi-state dynamic occupancy (multi-season) occupancy model using the second (conditional binomial) parameterisation in PRESENCE.}
\usage{
occMod_DO_ms2(psi = call(), psi.cov = data$unitcov, r = call(),
  r.cov = data$unitcov, p = call(), p.cov = data$survcov,
  delta = call(), delta.cov = data$survcov, modname = NULL,
  paoname = NULL, outfile, model = 6100, fixed = NULL, initvals = NULL,
  data, conf, miscopts)
}
\arguments{
\item{psi}{the right-hand side of the formula for the model to fit for occupancy probability, both first year and dynamic occupancy. The terms \code{SEASON}, \code{DYN} and \code{PREV_STATE} can each be used without having been defined in \code{psi.cov}. \code{SEASON} allows for a seasonal effect, \code{DYN} enables dynamic occupancy probabilities to be different from first year occupancy, and \code{PREV_STATE} allows the probability to be different depending on the state in the previous season.}

\item{psi.cov}{a data frame containing the unit-specific covariates to use for the occupancy component of the model.}

\item{r}{the right-hand side of the formula for the model to fit for the probability of being in the second state, conditional on the unit being occupied. The terms \code{SEASON}, \code{DYN} and \code{PREV_STATE} can be used as above with the same effect.}

\item{r.cov}{a data frame containing the unit-specific covariates to use for the probability of being in the second state, conditional on the unit being occupied part of the model.}

\item{p}{the right-hand side of the formula for the model to fit for detection probability. The terms \code{SEASON} and \code{STATE} can each be used without having been defined in \code{p.cov}. \code{SEASON} allows for a seasonal effect, and \code{STATE} allows the probability of detection to be different depending on the state in the current season.}

\item{p.cov}{a data frame containing the survey-specific covariates to use for the detection component of the model.}

\item{delta}{the right-hand side of the formula for the model to fit for the probability of detecting the second state in a survey, conditional on the species being detected in the survey. The terms \code{SEASON} and \code{STATE} can each be used as above to the same effect.}

\item{delta.cov}{a data frame containing the survey-specific covariates to use for the probability of detecting the second state in a survey, conditional on the species being detected in the survey.}

\item{modname}{(optional) a string containing the model name}

\item{paoname}{(optional) a string containing the filename for the temporary PRESENCE data file.}

\item{outfile}{name for output file (use outfile='modname') for outfile named via model name}

\item{model}{the PRESENCE model code. DO NOT CHANGE.}

\item{fixed}{a single-column matrix containing values for real parameters to be fixed at. \code{rownnames(fixed)} should contain the index of the real parameters to be fixed.}

\item{initvals}{initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.}

\item{data}{the \code{pao} data object containing the detection data and other information.}

\item{conf}{level for confidence interval (may be vector valued).}

\item{miscopts}{see \code{\link{occMod}}

NOTE THAT THIS FUNCTION HAS NOT BEEN EXTENSIVELY TESTED SO EXACT IMPLEMENTATION MAY CHANGE.}
}
\value{
returns a list of class \code{"occMod"} and \code{"do.1"}.

\code{occMod$beta} contains the objects:
 \item{psi}{estimated logistic regression coefficients and standard errors for probability of occurrence in the first year.}
 \item{psi.VC}{variance-covariance matrix for logistic regression coefficients for probability of occurrence in the first year.}
 \item{p}{estimated logistic regression coefficients and standard errors for probability of detection.}
 \item{p.VC}{variance-covariance matrix for logistic regression coefficients for probability of detection.}
 \item{VC}{the full variance-covariance matrix for all logistic regression coefficients.}

 \code{occMod$real} contains the objects:
 \item{psi}{estimated probabilities of occurrence for each sampling unit, along with standard errors and limits of 95\% confidence interval. Estimates are provided for the first season (calculated directly from the estimated \code{beta} parameters). Estimates for later seasons are provided as derived parameters. The season for which an estimate applies to can be identified from the final number of the row names (\code{rownames(psi)}).}
 \item{p}{estimated probabilities of detection for each survey, along with standard errors and limits of 95\% confidence interval.}

 \code{occMod$derived} contains the objects:
 \item{psi}{estimated probabilities of occurrence for each sampling unit for second season onwards, along with standard errors and limits of 95\% confidence interval. The season for which an estimate applies to can be identified from the final number of the row names (\code{rownames(psi)}).}
}
\description{
This is not intended for direct use, but instead the \code{\link{occMod}} function should be used with \code{type="do.ms.2"}.
}
\seealso{
\code{\link{occMod}}
}
\author{
Darryl MacKenzie
}
