% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ.mod.so.cd.R
\name{occMod_SO_cd}
\alias{occMod_SO_cd}
\title{Fit static occupancy, correlated detections model}
\usage{
occMod_SO_cd(psi = call(), psi.cov = data$unitcov, theta = call(),
  theta.cov = data$unitcov, p = call(), p.cov = data$survcov,
  th0pi = call(), th0pi.cov = data$unitcov, modname = NULL,
  paoname = NULL, outfile, model = 103, fixed = NULL, initvals = NULL,
  data, conf, miscopts = "")
}
\arguments{
\item{psi}{the right-hand side of the formula for the model to fit for occupancy probability (eg., ~1).}

\item{psi.cov}{a data frame containing the unit-specific covariates to use for the occupancy component of the model.}

\item{theta}{the right-hand side of the formula for the model to fit for secondary-scale occupancy probability.
(eg., ~PRIME -> th0 != th1, ~SURVEY -> th0(1)=th1(1), th0(2)=th1(2),...)}

\item{theta.cov}{a data frame containing the secondary-scale specific covariates to use for the secondary-scaleoccupancy component of the model.}

\item{p}{the right-hand side of the formula for the model to fit for detection probability.}

\item{p.cov}{a data frame containing the survey-specific covariates to use for the detection component of the model.}

\item{th0pi}{the right-hand side of the formula for the model to fit for local use before 1st segment.}

\item{th0pi.cov}{a data frame containing the unit-specific covariates to use for th0pi.}

\item{modname}{(optional) a string containing the model name}

\item{paoname}{(optional) a string containing the filename for the temporary PRESENCE data file.}

\item{outfile}{name for output file (use outfile='modname') for outfile named via model name}

\item{model}{the PRESENCE model code. DO NOT CHANGE.}

\item{fixed}{a single-column matrix containing values for real parameters to be fixed at. \code{rownnames(fixed)} should contain the index of the real parameters to be fixed.}

\item{initvals}{initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.}

\item{data}{the \code{pao} data object containing the detection data and other information.}

\item{conf}{level for confidence interval (may be vector valued).}

\item{miscopts}{see \code{\link{occMod}}}
}
\value{
list of class \code{"occMod"} and \code{"soCd"}.

\code{occMod$beta} contains the objects:
  \item{psi}{estimated logistic regression coefficients and standard errors for probability of occurrence.}
  \item{psi.VC}{variance-covariance matrix for logistic regression coefficients for probability of occurrence.}
  \item{theta}{estimated logistic regression coefficients and standard errors for probability of occurrence at the secondary scale.}
  \item{theta.VC}{variance-covariance matrix for logistic regression coefficients for probability of occurrence at the secondary scale.}
  \item{p}{estimated logistic regression coefficients and standard errors for probability of detection.}
  \item{p.VC}{variance-covariance matrix for logistic regression coefficients for probability of detection.}
  \item{th0pi}{estimated logistic regression coefficients and standard errors for probability of local-use before 1st segment.}
  \item{th0pi.VC}{variance-covariance matrix for logistic regression coefficients for th0pi.}
  \item{VC}{the full variance-covariance matrix for all logistic regression coefficients.}

\code{occMod$real} contains the objects:
  \item{psi}{estimated probabilities of occurrence for each sampling unit, along with standard errors and limits of 95\% confidence interval.}
  \item{theta}{estimated probabilities of occurrence at the secondary scale for each sampling unit, along with standard errors and limits of 95\% confidence interval.}
  \item{p}{estimated probabilities of detection for each survey, along with standard errors and limits of 95\% confidence interval.}
  \item{psi_c}{estimated probabilities of occurrence given the detection history for each sampling unit, along with standard errors and limits of 95\% confidence interval. Will be \code{=1} for any unit where the species was detected at least once.}
}
\description{
This is not intended for direct use, but instead the \code{\link{occMod}} function should be used with \code{type="so.cd"}.
}
\examples{
#   example 1 - read previously created input file and run a model
data=readPao('genpres.pao')  #  read simulated data from program GENPRES
mod1=occMod(model=list(psi~1, theta~PRIME, p~1, th0pi~1), data=data, type="so.cd")

#   example 2 - simulate data and run a mode
sim_corr_det_data <- function(sites=500,surveys=9,psi=.9,th=c(.2,.8),p=.6,th0pi=0) {
  # simulates single-season correlated detections data
  h=matrix(0,sites,surveys); px=c(0,p) # px=detection prob for each state (local-unocc, local-occ)
  occ=0+(runif(sites)<psi);            # randomly assign occupancy state of each site (0=unocc, 1=occ)
  locc=occ*(runif(sites)<th0pi);       # randomly assign local-occ state before 1st segment of trail
  for (i in 1:surveys) {
     locc=occ*(runif(sites)<th[locc+1]);   # local-occ for survey i depends on previous local-occ
     h[,i]=locc*(runif(sites)<px[locc+1])  # detection depends on local-occ state (p=0 if not local-occ)
  }
return(h)
}
x=sim_corr_det_data(sites=1000,surveys=15)
data=createPao(x)
mod1=occMod(model=list(psi~1,       # constant psi
                     theta~PRIME, # theta not equal theta'
                     p~1,         # constant detection
                     th0pi~1      # constant prop unocc before 1st segment
), data=data, type="so.cd")  #  run ss-cd model
print_one_site_estimates(mod1,site=1)

}
