% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{print_one_site_estimates}
\alias{print_one_site_estimates}
\title{Print real param estimates for i-th site only
\code{print_one_site_estimates} Prints real param estimates for ith site only}
\usage{
print_one_site_estimates(mod, site = 1)
}
\arguments{
\item{mod}{model output from \code{\link{occMod}} function.}

\item{site}{site number(s) to print (default=1)}
}
\value{
nothing
}
\description{
Print real param estimates for i-th site only
\code{print_one_site_estimates} Prints real param estimates for ith site only
}
