% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{readPao}
\alias{readPao}
\title{Read a pao data object.}
\usage{
readPao(f)
}
\arguments{
\item{f}{A string specifying the name of the PRESENCE data file.}
}
\value{
list of class "\code{pao}" containing the data required to fit an
 occupancy model using \code{\link{occMod}}.
 The list contains the following objects:

  \item{nunits}{number of units or sampling units, i.e., \code{nrow(data)}.}
  \item{nsurveys}{total number of surveys, i.e., \code{ncol(data)}.}
  \item{nseasons}{number of seasons of data.}
  \item{det.data}{detection/nondetection data, i.e., \code{data}.}
  \item{unitcov}{see above. If no covariates are supplied a vector of 1's named TEMP is added.}
  \item{survcov}{see above. A factor called SURVEY is automatically added to allow survey-specific detection probabilities (but equal across units).}
  \item{nsurveyseason}{see above.}
  \item{title}{see above.}
  \item{unitnames}{see above.}
  \item{surveynames}{a vector of strings providing labels for the survey occasions. Of little practical consequence.}
  \item{paoname}{see above. Equals \code{f} for \code{createPao}.}
}
\description{
\code{readPao} Creates a pao data object from a PRESENCE data file.
}
\examples{
readPao(system.file('extdata/skinks.pao',package='RPresence'))

}
\seealso{
\code{\link{occMod}}
}
\author{
Darryl MacKenzie and Jim Hines
}
