% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_residual_functions.R
\name{residuals.occMod}
\alias{residuals.occMod}
\title{A function produce the Dunn-Smyth residuals.}
\usage{
\method{residuals}{occMod}(object, type = "DS", ...)
}
\arguments{
\item{object}{the occMod object which must contain the following:}
}
\value{
a list with two components:
\item{det}{Dunn-Smyth residuals for detection, based on cdf of #detections}
\item{occ}{Dunn-Smyth residuals for occupancy, based on cdf of binary detected/not)}
}
\description{
\code{residuals.occMod} produce the Dunn-Smyth residuals.
}
\details{
\itemize{
   \item{$data$psi$est} {fitted probabilities of occupancy}
   \item{$data$p$est} {fitted probabilities of detection.}
}
$data$det.data, binary matrix of detections, sites in rows and visits in columns. NA for missing values.
}
