% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{runPresence}
\alias{runPresence}
\title{call PRESENCE to produce model estimates (replaces "write_dm_and_run")}
\usage{
runPresence(a, dm, model, modname = "m1", fixed = NULL, initvals = NULL,
  outfile, miscopts = "")
}
\arguments{
\item{a}{pao object}

\item{dm}{list of design matrices (should be 6 in list)}

\item{model}{model number (not used anymore)}

\item{modname}{name of model (eg., "psi(.)p(.)")}

\item{fixed}{data.frame of fixed values (e.g., fixed$params<-c("p[1]","p[3]"),fixed$value<-0)}

\item{initvals}{vector of beta initial values}

\item{outfile}{name for output file (use outfile='modname') for outfile named via model name}

\item{miscopts}{options (VCoutput, quiet, chsq,...)}
}
\value{
list(LL=log-likelihood, beta=beta estimates, betaVC=var-cov matrix of betas).
}
\description{
\code{runPresence} calls PRESENCE to produce model estimates.
}
\details{
Input is passed
to PRESENCE via arguments (no pao or design matrix input files needed).  There are
three choices for output:

\itemize{
\item outfile=NULL - no output file saved, results in R variables only,
\item outfile="modname" - output file saved with name of model and ".out" appended,
\item outfile="xyz" - output file saved with user-specified name, (eg., "xyz.out")
}
The first option is best for running simulations of a model using "Snowfall"
for multi-tasking.  The 2nd option is useful when you would like to find
the full text output of a model from Windows explorer.  The 3rd option allows
you to specify any output file name.
}
\examples{
runPresence("test.pao",list(dm1,dm2,dm3,dm4,dm5,dm6),100,"psi(.)p(.)")

}
\seealso{
\code{\link{occMod}}
}
\author{
Darryl MacKenzie and Jim Hines
}
