% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{summary.occMod}
\alias{summary.occMod}
\title{Summary of occupancy model}
\usage{
\method{summary}{occMod}(object, ...)
}
\arguments{
\item{object}{an object of class \code{occMod} which is the result from fitting a model using the \code{\link{occMod}} function.}

\item{...}{possible additional arguments}
}
\value{
Provides a brief summary of the fitted model, including model name, AIC value and number of parameters.
}
\description{
Provides a brief summary of an object of class \code{occMod}, which is the result from fitting amodel using the \code{\link{occMod}} function.
}
\author{
Darryl MacKenzie
}
\keyword{htest}
