% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{summary.pao}
\alias{summary.pao}
\title{Summary of occupancy data object}
\usage{
\method{summary}{pao}(object)
}
\arguments{
\item{object}{an object of class \code{pao} created by \code{readPao} or \code{createPao}.}
}
\value{
Provides a brief summary of the input object, including number of sites, surveys.
}
\description{
Provides a brief summary of an object of class \code{pao}, which is the input data for occupancy analysis.
}
\author{
Jim Hines
}
\keyword{pao}
