% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-PRES.r
\name{writePao}
\alias{writePao}
\title{Write a pao data object to disk file.}
\usage{
writePao(pao)
}
\arguments{
\item{pao}{A pao object created with \code{readPao} or \code{createPao}.}
}
\value{
name of file written.
}
\description{
\code{writePao} Writes a pao data object to disk.
}
\examples{
\dontrun{
writePao(paodata)
}

}
\seealso{
\code{\link{occMod}}
}
\author{
Darryl MacKenzie and Jim Hines
}
